requirements.txt
colorama
termcolor
discord.py[voice]
PyNaCl